transitioned to GLFW3, as I like the callback system better than the Event States of SDL2

+++ src/utils/interval.h
@@ -3,11 +3,11 @@
 namespace Glow {namespace utils{
 
     struct Interval {
-        Interval (long interval, void (*callback)(), int id) : interval_(interval),
+        Interval (double interval, void (*callback)(), int id) : interval_(interval),
             delay_(interval), callback_(callback), id_(id){};
 
-        long interval_;
-        long delay_;
+        double interval_;
+        double delay_;
         void (*callback_)();
         int id_;
     };
